daten/karlsruhe/wahldaten-scraper/requirements.txt
beautifulsoup4==4.6.0
backports.csv==1.0.5
